[audio] Move WhisperTranslator to audio extension
